==> Python Notebooks/1_Create_Train_Test_Cohorts.ipynb <==
# Create Train and Test Cohorts

## Load libraries and enable Pyspark and PyArrow to allow for distributed processing

import os
os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-8-openjdk-amd64"

import numpy as np
import pandas as pd
import csv
import pickle
import time
import math
import collections
import os
from tqdm import tqdm
import pyspark
from pyspark.sql import *
from pyspark.sql.functions import *
from pyspark import SparkContext, SparkConf
import pyspark.sql.functions as f

# create the session
conf = SparkConf().set("spark.ui.port", "4050")
conf = (conf.setMaster('local[*]')
        .set('spark.executor.memory', '35G')
        .set('spark.driver.memory', '35G')
        .set('spark.driver.maxResultSize', '35G'))
# create the context
sc = pyspark.SparkContext(conf=conf)
spark = SparkSession.builder.getOrCreate()
spark.conf.set("spark.sql.execution.arrow.enabled", "true")
# arrow enabling is what makes the conversion from pandas to spark dataframe really fast
sc._conf.get('spark.driver.memory')

## Read in raw data and select for diagnosis code columns (16-41) and demographic information (0-8)

dirPath = '/home/ubuntu/BioMedProject/Data/'
patientDataPath = dirPath + 'patientData.csv' # file was renamed from 'B220_SAA_v1.csv'

begin = time.time()
df = spark.read.load(patientDataPath, format="csv", sep=",", inferSchema="true", header="true")
df = df.select(df.columns[0:8] + df.columns[16:41])
df = df.drop('Type')
end = time.time()
print("Reading csv took: {} seconds".format(end - begin))

## Create 30-day readmission labels based on immediate future visit for each patient

from pyspark.sql.window import Window
from pyspark.sql.functions import lit

begin = time.time()
my_window = Window.partitionBy().orderBy("ID")

df = df.withColumn("next_Date", lead(df.Date).over(my_window))
df = df.withColumn("next_Id", lead(df.ID).over(my_window))
df = df.withColumn("decFirst", lit("2018-12-01"))
df = df.withColumn("daysUntilDecFirst", datediff(col("decFirst"), col("Date")).alias("finalVisitDiff"))
df = df.withColumn("datediff", datediff(col("next_Date"), col("Date")).alias("datediff"))
end = time.time()
print("Cell took {}".format(end - begin))

begin = time.time()
df = df.withColumn("label", when((f.col("ID") != f.col("next_Id")) & (f.col("daysUntilDecFirst") >= 0), 0)\
                   .when((f.col("ID") != f.col("next_Id")) & (f.col("daysUntilDecFirst") < 0), 2)\
                   .when(col("datediff") > 30, 0)\
                   .when(col("datediff") <= 30, 1)\
                   .otherwise(2))
end = time.time()
print("Cell took {}".format(end - begin))

df = df.filter(f.col("label") != 2)

total_count = df.count()
print("Total number of visits: {}".format(total_count))

df = df.drop("next_Date", "next_Id", "decFirst", "daysUntilDecFirst", "datediff")
df.printSchema()

# COUNTS NUMBER OF VISITS FILTERED OUT:
# df.filter(f.col("daysUntilDecFirst") < 0).select(countDistinct("ID")).show()

Visits to keep: 27502639

Visits filtered out: 475293

Total visits: 27977932

## Shuffle the data at random

from pyspark.sql.functions import monotonically_increasing_id 

df = df.orderBy(rand(seed=1234))
df = df.withColumn("index", monotonically_increasing_id())

assert df.count()== total_count

## Create an 80-20 train test split

split = int(total_count * 0.8)
print("Total count: {}".format(total_count))
print("Splitting at: {}".format(split))

train = df.filter(col("index").between(0, split)).drop("index") # repartitin or coalesce after filtering -- 1GB per partition rule
test = df.filter(col("index").between(split + 1, total_count)).drop("index")

print("Train size: {}".format(train.count()))
print("Test size: {}".format(test.count()))

## Save train and test datasets

begin = time.time()
train.write.format('csv').option('header', True).option('sep',',').mode('overwrite').save(dirPath + 'train')
end = time.time()
print("Saving train file took {} seconds".format(end - begin))

begin = time.time()
test.write.format('csv').option('header', True).option('sep',',').mode('overwrite').save(dirPath + 'test')
end = time.time()
print("Saving test file took {} seconds".format(end - begin))

## Load from saved files

begin = time.time()
train_loaded = spark.read.format("csv").load(dirPath + "train",
                     sep=",", inferSchema="true", header="true")
end = time.time()
print("Cell took {} seconds".format(end - begin))

begin = time.time()
test_loaded = spark.read.load(dirPath + "test",
                     format="csv", sep=",", inferSchema="true", header="true")
end = time.time()
print("Cell took {} seconds".format(end - begin))

## Assert no errors during saving

assert train.count() == train_loaded.count()
assert test.count() == test_loaded.count()
