==> Python Notebooks/4_Run_Models.ipynb <==
# Run ML Models

import pandas as pd
import numpy as np
import time
import random
import pickle
import csv
import collections
from tqdm import tqdm

import pyspark
from pyspark.sql import *
from pyspark.sql.functions import *
from pyspark.sql.window import Window
from pyspark.ml.feature import VectorAssembler
from pyspark import SparkContext, SparkConf
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.ml.evaluation import BinaryClassificationEvaluator

# create the session
conf = SparkConf().set("spark.ui.port", "4050")
conf = (conf.setMaster('local[*]')
        .set('spark.executor.memory', '50G')
        .set('spark.driver.memory', '100G')
        .set('spark.driver.maxResultSize', '50G'))
# create the context
sc = pyspark.SparkContext(conf=conf)
spark = SparkSession.builder.getOrCreate()
spark.conf.set("spark.sql.execution.arrow.enabled", "true")
# arrow enabling is what makes the conversion from pandas to spark dataframe really fast
sc._conf.get('spark.driver.memory')

spark

## Linear SVC

def trainSVMModel(train, test, model_id, embed, maxIter, reg, txt_filepath, csv_filepath):
    # Train a Linear SVC model.
    start = time.time()
    svm = LinearSVC(labelCol="Label", featuresCol="features")
    svm.setMaxIter(maxIter)
    svm.setRegParam(reg)
    svm_model = svm.fit(train)
    end = time.time()
    
    # Save model
    model_path = dirPath + f"ModelResults/LinearSVC_EMBED{embed}_MAXITER{maxIter}_REG{reg}_AgeModif.model"
    svm_model.write().overwrite().save(model_path)
    
    # Predictions
    predictions_train = svm_model.transform(train)
    predictions_test = svm_model.transform(test)
    
    # Accuracy and AUROC
    train_accuracy = computeAccuracy(predictions_train)
    test_accuracy = computeAccuracy(predictions_test)
    train_auroc = computeAUROC(predictions_train)
    test_auroc = computeAUROC(predictions_test)
    
    timeElapsed = np.round((end-start)/60, 2)
    
    # Save results to readable text file 
    txt_filepath.write(f"{model_id}) SVM MODEL: {embed} embed length, {maxIter} max iter, {reg} reg\n")
    txt_filepath.write(f"\t Time: {timeElapsed} minutes\n")
    txt_filepath.write(f"\t TRAIN Accuracy: {train_accuracy}\n")
    txt_filepath.write(f"\t TRAIN AUROC:    {train_auroc}\n")
    txt_filepath.write(f"\t TEST  Accuracy: {test_accuracy}\n")
    txt_filepath.write(f"\t TEST  AUROC:    {test_auroc}\n")
    txt_filepath.write("\n")
    
    # Save results to csv file
    csv_result = f"{model_id}, {embed}, {maxIter}, {reg}, {timeElapsed}, {train_accuracy}, {train_auroc}, {test_accuracy}, {test_auroc}\n"
    csv_filepath.write(csv_result)
    

def computeAccuracy(predictions):
    evaluator = MulticlassClassificationEvaluator(
        labelCol="Label", predictionCol="prediction", metricName="accuracy")
    accuracy = evaluator.evaluate(predictions)
    return accuracy

def computeAUROC(predictions):
    evaluator = BinaryClassificationEvaluator(labelCol="Label", rawPredictionCol="prediction", metricName="areaUnderROC")
    auroc = evaluator.evaluate(predictions)
    return auroc

from pyspark.ml.classification import LinearSVC, LinearSVCModel

dirPath = '/home/ubuntu/BioMedProject/Data/'
resultsFilePath = open(dirPath + 'SVM_Results_Readable.txt', "w")
csvFilePath = open(dirPath + 'SVM_Results.csv', "w")

EMBED_DIM =  [16, 32, 64, 128] #, 256]
MAX_ITER = [5, 10]
REG_PARAM = [0.0, 0.01, 0.1, 0.3]

model_id = 0
for embed in EMBED_DIM:
    
    trainFilepath = dirPath + f'ModelData/modelReadyTrain_{embed}.parquet'
    testFilepath = dirPath + f'ModelData/modelReadyTest_{embed}.parquet'
    df_test = spark.read.load(testFilepath, format="parquet", inferSchema="true", header="true")
    df_train = spark.read.load(trainFilepath, format="parquet", inferSchema="true", header="true")
    
    for maxIter in MAX_ITER:
        for reg in REG_PARAM:
            model_id += 1
            t1 = time.time()
            trainSVMModel(df_train, df_test, model_id, embed, maxIter, reg, resultsFilePath, csvFilePath)
            t2 = time.time()
            print(f"id: {model_id}, {embed} embed, {maxIter} max iter, {reg} reg, {(t2 - t1)/60} min")

csvFilePath.close()
resultsFilePath.close()

!cat 'Data/SVM_Results_Readable.txt'

## GBT classifier

def trainGBTModel(train, test, model_id, embed, maxDepth, txt_filepath, csv_filepath):
    # Train a GBT model.
    start = time.time()
    svm = GBTClassifier(labelCol="Label", featuresCol="features")
    svm.setMaxDepth(maxDepth)
    svm_model = svm.fit(train)
    end = time.time()
    
    # Save model
    model_path = dirPath + f"ModelResults/GBT_EMBED{embed}_MAXDEPTH{maxDepth}.model"
    svm_model.write().overwrite().save(model_path)
    
    # Predictions
    predictions_train = svm_model.transform(train)
    predictions_test = svm_model.transform(test)
    
    # Accuracy and AUROC
    train_accuracy = computeAccuracy(predictions_train)
    test_accuracy = computeAccuracy(predictions_test)
    train_auroc = computeAUROC(predictions_train)
    test_auroc = computeAUROC(predictions_test)
    
    timeElapsed = np.round((end-start)/60, 2)
    
    # Save results to readable text file 
    txt_filepath.write(f"{model_id}) GBT MODEL: {embed} embed length, {maxDepth} max depth,\n")
    txt_filepath.write(f"\t Time: {timeElapsed} minutes\n")
    txt_filepath.write(f"\t TRAIN Accuracy: {train_accuracy}\n")
    txt_filepath.write(f"\t TRAIN AUROC:    {train_auroc}\n")
    txt_filepath.write(f"\t TEST  Accuracy: {test_accuracy}\n")
    txt_filepath.write(f"\t TEST  AUROC:    {test_auroc}\n")
    txt_filepath.write("\n")
    
    # Save results to csv file
    csv_result = f"{model_id}, {embed}, {maxDepth}, {timeElapsed}, {train_accuracy}, {train_auroc}, {test_accuracy}, {test_auroc}\n"
    csv_filepath.write(csv_result)
    

def computeAccuracy(predictions):
    evaluator = MulticlassClassificationEvaluator(
        labelCol="Label", predictionCol="prediction", metricName="accuracy")
    accuracy = evaluator.evaluate(predictions)
    return accuracy

def computeAUROC(predictions):
    evaluator = BinaryClassificationEvaluator(labelCol="Label", rawPredictionCol="prediction", metricName="areaUnderROC")
    auroc = evaluator.evaluate(predictions)
    return auroc

from pyspark.ml.classification import GBTClassifier, GBTClassificationModel

dirPath = '/home/ubuntu/BioMedProject/Data/'
resultsFilePath = open(dirPath + 'GBT_Results_Readable.txt', "a") # append mode since models 1-5 were already run
csvFilePath = open(dirPath + 'GBT_Results.csv', "a")

EMBED_DIM =  [16, 32, 64, 128] #, 256]
MAXDEPTH_PARAM = [5, 10] #[15, 25, 50, 100]

model_id = 1
for embed in EMBED_DIM:
    
    trainFilepath = dirPath + f'ModelData/modelReadyTrain_{embed}.parquet'
    testFilepath = dirPath + f'ModelData/modelReadyTest_{embed}.parquet'
    df_test = spark.read.load(testFilepath, format="parquet", inferSchema="true", header="true")
    df_train = spark.read.load(trainFilepath, format="parquet", inferSchema="true", header="true")
    
    for maxDepth in MAXDEPTH_PARAM:
        model_id += 1
        if model_id < 6: continue
        trainGBTModel(df_train, df_test, model_id, embed, maxDepth, resultsFilePath, csvFilePath)
        print(f"{model_id}, {embed}, {maxDepth}")

csvFilePath.close()
resultsFilePath.close()


!cat 'Data/GBT_Results_Readable.txt'

## Random Forest Classifer

def trainRFModel(train, test, model_id, embed, maxDepth, numTrees, txt_filepath, csv_filepath):
    # Train a RF model.
    start = time.time()
    rf = RandomForestClassifier(labelCol="Label", featuresCol="features")
    rf.setMaxDepth(maxDepth)
    rf.setNumTrees(numTrees)
    rf_model = rf.fit(train)
    end = time.time()
    
    # Save model
    model_path = dirPath + f"ModelResults/RF_EMBED{embed}_MAXDEPTH{maxDepth}_NUMTREES{numTrees}.model"
    rf_model.write().overwrite().save(model_path)
    
    # Predictions
    predictions_train = rf_model.transform(train)
    predictions_test = rf_model.transform(test)
    
    # Accuracy and AUROC
    train_accuracy = computeAccuracy(predictions_train)
    test_accuracy = computeAccuracy(predictions_test)
    train_auroc = computeAUROC(predictions_train)
    test_auroc = computeAUROC(predictions_test)
    
    timeElapsed = np.round((end-start)/60, 2)
    
    # Save results to readable text file 
    txt_filepath.write(f"{model_id}) RF MODEL: {embed} embed length, {maxDepth} max depth, {numTrees} num trees\n")
    txt_filepath.write(f"\t Time: {timeElapsed} minutes\n")
    txt_filepath.write(f"\t TRAIN Accuracy: {train_accuracy}\n")
    txt_filepath.write(f"\t TRAIN AUROC:    {train_auroc}\n")
    txt_filepath.write(f"\t TEST  Accuracy: {test_accuracy}\n")
    txt_filepath.write(f"\t TEST  AUROC:    {test_auroc}\n")
    txt_filepath.write("\n")
    
    # Save results to csv file
    csv_result = f"{model_id}, {embed}, {maxDepth}, {numTrees}, {timeElapsed}, {train_accuracy}, {train_auroc}, {test_accuracy}, {test_auroc}\n"
    csv_filepath.write(csv_result)
    

def computeAccuracy(predictions):
    evaluator = MulticlassClassificationEvaluator(
        labelCol="Label", predictionCol="prediction", metricName="accuracy")
    accuracy = evaluator.evaluate(predictions)
    return accuracy

def computeAUROC(predictions):
    evaluator = BinaryClassificationEvaluator(labelCol="Label", rawPredictionCol="prediction", metricName="areaUnderROC")
    auroc = evaluator.evaluate(predictions)
    return auroc

from pyspark.ml.classification import RandomForestClassifier, RandomForestClassificationModel

dirPath = '/home/ubuntu/BioMedProject/Data/'
resultsFilePath = open(dirPath + 'RF_Results_Readable.txt', "a")
csvFilePath = open(dirPath + 'RF_Results.csv', "a")

EMBED_DIM =  [16, 32, 64, 128] #, 256]
MAXDEPTH_PARAM = [15]
NUMBER_TREES = [20]

model_id = 4
for embed in EMBED_DIM:
    
    trainFilepath = dirPath + f'ModelData/modelReadyTrain_{embed}.parquet'
    testFilepath = dirPath + f'ModelData/modelReadyTest_{embed}.parquet'
    df_test = spark.read.load(testFilepath, format="parquet", inferSchema="true", header="true")
    df_train = spark.read.load(trainFilepath, format="parquet", inferSchema="true", header="true")
    
    for maxDepth in MAXDEPTH_PARAM:
        for numTrees in NUMBER_TREES:
            model_id += 1
            t1 = time.time()
            trainRFModel(df_train, df_test, model_id, embed, maxDepth, numTrees, resultsFilePath, csvFilePath)
            t2 = time.time()
            print(f"id: {model_id}, {embed} embed, {maxDepth} depth, {numTrees} trees, {(t2 - t1)/60} min")

csvFilePath.close()
resultsFilePath.close()

!cat 'Data/RF_Results_Readable.txt'

## Logistic Regression

def trainLRModel(train, test, model_id, embed, maxIter, reg, txt_filepath, csv_filepath):
    # Train a Logistic Regression model.
    start = time.time()
    lr = LogisticRegression(labelCol="Label", featuresCol="features", maxIter=maxIter, regParam=reg)
    lr_model = lr.fit(train)
    end = time.time()
    
    # Save model
    model_path = dirPath + f"ModelResults/LogisticReg_EMBED{embed}_MAXITER{maxIter}_REG{reg}.model"
    lr_model.write().overwrite().save(model_path)
    
    # Predictions
    predictions_train = lr_model.transform(train)
    predictions_test = lr_model.transform(test)
    
    # Accuracy and AUROC
    train_accuracy = computeAccuracy(predictions_train)
    test_accuracy = computeAccuracy(predictions_test)
    train_auroc = computeAUROC(predictions_train)
    test_auroc = computeAUROC(predictions_test)
    
    timeElapsed = np.round((end-start)/60, 2)
    
    # Save results to readable text file 
    txt_filepath.write(f"{model_id}) LR MODEL: {embed} embed length, {maxIter} max iter, {reg} reg\n")
    txt_filepath.write(f"\t Time: {timeElapsed} minutes\n")
    txt_filepath.write(f"\t TRAIN Accuracy: {train_accuracy}\n")
    txt_filepath.write(f"\t TRAIN AUROC:    {train_auroc}\n")
    txt_filepath.write(f"\t TEST  Accuracy: {test_accuracy}\n")
    txt_filepath.write(f"\t TEST  AUROC:    {test_auroc}\n")
    txt_filepath.write("\n")
    
    # Save results to csv file
    csv_result = f"{model_id}, {embed}, {maxIter}, {reg}, {timeElapsed}, {train_accuracy}, {train_auroc}, {test_accuracy}, {test_auroc}\n"
    csv_filepath.write(csv_result)   

def computeAccuracy(predictions):
    evaluator = MulticlassClassificationEvaluator(
        labelCol="Label", predictionCol="prediction", metricName="accuracy")
    accuracy = evaluator.evaluate(predictions)
    return accuracy

def computeAUROC(predictions):
    evaluator = BinaryClassificationEvaluator(labelCol="Label", rawPredictionCol="prediction", metricName="areaUnderROC")
    auroc = evaluator.evaluate(predictions)
    return auroc

from pyspark.ml.classification import LogisticRegression, LogisticRegressionModel

dirPath = '/home/ubuntu/BioMedProject/Data/'
resultsFilePath = open(dirPath + 'LR_Results_Readable.txt', "w")
csvFilePath = open(dirPath + 'LR_Results.csv', "w")

EMBED_DIM =  [16, 32, 64, 128] #, 256]
MAX_ITER = [5, 10]
REG_PARAM = [0.0, 0.01, 0.1, 0.3]

model_id = 0
for embed in EMBED_DIM:
    
    trainFilepath = dirPath + f'ModelData/modelReadyTrain_{embed}.parquet'
    testFilepath = dirPath + f'ModelData/modelReadyTest_{embed}.parquet'
    df_test = spark.read.load(testFilepath, format="parquet", inferSchema="true", header="true")
    df_train = spark.read.load(trainFilepath, format="parquet", inferSchema="true", header="true")
    
    for maxIter in MAX_ITER:
        for reg in REG_PARAM:
            model_id += 1
            t1 = time.time()
            trainLRModel(df_train, df_test, model_id, embed, maxIter, reg, resultsFilePath, csvFilePath)
            t2 = time.time()
            print(f"id: {model_id}, {embed} embed, {maxIter} max iter, {reg} reg, {(t2 - t1)/60} min")

csvFilePath.close()
resultsFilePath.close()

# Results

## SVM Classifier

headers = ["Model ID", "Embed Length", "Max Iter", "Reg", "Time Elapsed (min)", "Train Acc", "Train AUROC", "Test Acc", "Test AUROC"]
results = pd.read_csv(dirPath + 'SVM_Results.csv', header=None, names = headers)
results.sort_values(by = "Test AUROC", ascending = False)

## Random Forest Results

headers = ["Model ID", "Embed Length", "Max Depth", "Num Trees", "Time Elapsed (min)", "Train Acc", "Train AUROC", "Test Acc", "Test AUROC"]
results = pd.read_csv(dirPath + 'RF_Results.csv', header=None, names = headers)
results.sort_values(by = "Test AUROC", ascending = False)

## GBT Classifier Result

headers = ["Model ID", "Embed Length", "Max Depth", "Time Elapsed (min)", "Train Acc", "Train AUROC", "Test Acc", "Test AUROC"]
results = pd.read_csv(dirPath + 'GBT_Results.csv', header=None, names = headers)
results.sort_values(by = "Test AUROC", ascending = False)

## Logistic Regression Results

headers = ["Model ID", "Embed Length", "Max Iter", "Reg", "Time Elapsed (min)", "Train Acc", "Train AUROC", "Test Acc", "Test AUROC"]
results = pd.read_csv(dirPath + 'LR_Results.csv', header=None, names = headers)
results.sort_values(by = "Test AUROC", ascending = False)
