==> Python Notebooks/4a_Run_h2o_Models.ipynb <==
# H20 Models

import pandas as pd
import numpy as np
import time
import random
import pickle
import csv
import collections
from tqdm import tqdm

import pyspark
from pyspark.sql import *
from pyspark.sql.functions import *
from pyspark.sql.window import Window
from pyspark.ml.feature import VectorAssembler
from pyspark import SparkContext, SparkConf
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.ml.evaluation import BinaryClassificationEvaluator
from pyspark.ml.feature import Tokenizer, Word2Vec, Word2VecModel, OneHotEncoder, StringIndexer
from pyspark.ml import Pipeline
import pyspark.sql.functions as F
import pyspark.sql.types as T

# create the session
conf = SparkConf().set("spark.ui.port", "4050")
conf = (conf.setMaster('local[*]')
        .set('spark.executor.memory', '50G')
        .set('spark.driver.memory', '100G')
        .set('spark.driver.maxResultSize', '50G'))
# create the context
sc = pyspark.SparkContext(conf=conf)
spark = SparkSession.builder.getOrCreate()
spark.conf.set("spark.sql.execution.arrow.enabled", "true")
# arrow enabling is what makes the conversion from pandas to spark dataframe really fast
sc._conf.get('spark.driver.memory')

import h2o
from h2o.estimators import H2OXGBoostEstimator
h2o.init(max_mem_size = '300G')

## Load in Data

dirPath = '/home/ubuntu/BioMedProject/Data/'
trainCorpusFilename = dirPath + "ModelData/trainCorpusAllVisits.pkl"
testCorpusFilename = dirPath + "ModelData/testCorpusAllVisits.pkl"

# Load Embeddings
EMBED_LENGTH = 128 # 16, 32, 64, 128, 256

#embedFilePath = dirPath + f"W2V_Models/w2v_ICD_embed_{EMBED_LENGTH}.model"
embedFilePath = dirPath + f"W2V_Models/w2v_ICD_Combined_Visits_embed_{EMBED_LENGTH}.model"
loaded_model = Word2VecModel.load(embedFilePath)

# Read train/test datasets
train = spark.read.load(dirPath + "train",
                     format="csv", sep=",", inferSchema="true", header="true")
test = spark.read.load(dirPath + "test",
                     format="csv", sep=",", inferSchema="true", header="true")

# Read in train/test corpus
with open(trainCorpusFilename, 'rb') as handle:
    corpus_train = pickle.load(handle)
    
with open(testCorpusFilename, 'rb') as handle:
    corpus_test = pickle.load(handle)

print("{} million TRAIN visits".format(len(corpus_train)/1e6))
print("{} million TEST visits".format(len(corpus_test)/1e6))

## Prepare train/test ICD features

# Train ICD features
start = time.time()
sentences_train = pd.DataFrame(corpus_train, columns = ['sentences'])
sentences_train = spark.createDataFrame(sentences_train)

tokenizer = Tokenizer(inputCol="sentences", outputCol="tokens")
tokenized_sentences_train = tokenizer.transform(sentences_train).select("tokens")

to_array = F.udf(lambda v: v.toArray().tolist(), T.ArrayType(T.FloatType()))
features_train = loaded_model.transform(tokenized_sentences_train)
features_train = features_train.withColumn('features', to_array('features'))
features_train = features_train.select([F.col("features")[i] for i in range(EMBED_LENGTH)])
features_train = features_train.toPandas()

del sentences_train, tokenized_sentences_train
end = time.time()
print(f"Cell took {np.round(end-start, 2)} seconds")
features_train.head(5)

# Test ICD features
start = time.time()
sentences_test = pd.DataFrame(corpus_test, columns = ['sentences'])
sentences_test = spark.createDataFrame(sentences_test)

tokenizer = Tokenizer(inputCol="sentences", outputCol="tokens")
tokenized_sentences_test = tokenizer.transform(sentences_test).select("tokens")

to_array = F.udf(lambda v: v.toArray().tolist(), T.ArrayType(T.FloatType()))
features_test = loaded_model.transform(tokenized_sentences_test)
features_test = features_test.withColumn('features', to_array('features'))
features_test = features_test.select([F.col("features")[i] for i in range(EMBED_LENGTH)])
features_test = features_test.toPandas()

del sentences_test, tokenized_sentences_test
end = time.time()
print(f"Cell took {np.round(end-start, 2)} seconds")
features_test.head(5)

## Prepare demographic features

start = time.time()
demographic_train = train.select("Age", "Sex", "Race", "Label")
indexers = [StringIndexer(inputCol=column, outputCol=column+"_NUMERIC").fit(demographic_train) for column in ['Sex', 'Race']]
pipeline = Pipeline(stages=indexers)
demographic_train = pipeline.fit(demographic_train).transform(demographic_train)
demographic_train = demographic_train.drop("Sex", "Race")
demographic_train = demographic_train.withColumn("Sex", demographic_train["Sex_NUMERIC"].cast(T.IntegerType()))
demographic_train = demographic_train.withColumn("Race", demographic_train["Race_NUMERIC"].cast(T.IntegerType()))
demographic_train = demographic_train.drop("Sex_NUMERIC", "Race_NUMERIC")
demographic_train = demographic_train.withColumn("Age", (5/100*round(train["Age"] / 5 )).cast(T.FloatType()))
df_demographic_train = demographic_train.toPandas()
df_demographic_train = pd.get_dummies(df_demographic_train, columns=["Sex", "Race"])

del demographic_train
end = time.time()

print(f"Cell took {np.round(end-start, 2)} seconds")
df_demographic_train.head(5)

start = time.time()
demographic_test = test.select("Age", "Sex", "Race", "Label")
indexers = [StringIndexer(inputCol=column, outputCol=column+"_NUMERIC").fit(demographic_test) for column in ['Sex', 'Race']]
pipeline = Pipeline(stages=indexers)
demographic_test = pipeline.fit(demographic_test).transform(demographic_test)
demographic_test = demographic_test.drop("Sex", "Race")
demographic_test = demographic_test.withColumn("Sex", demographic_test["Sex_NUMERIC"].cast(T.IntegerType()))
demographic_test = demographic_test.withColumn("Race", demographic_test["Race_NUMERIC"].cast(T.IntegerType()))
demographic_test = demographic_test.drop("Sex_NUMERIC", "Race_NUMERIC")
demographic_test = demographic_test.withColumn("Age", (5/100*round(test["Age"] / 5 )).cast(T.FloatType()))
df_demographic_test = demographic_test.toPandas()
df_demographic_test = pd.get_dummies(df_demographic_test, columns=["Sex", "Race"])

del demographic_test
end = time.time()

print(f"Cell took {np.round(end-start, 2)} seconds")
df_demographic_test.head(5)

## Concentate ICD and Demographic features

df_train = pd.concat([df_demographic_train, features_train], axis=1)
df_test = pd.concat([df_demographic_test, features_test], axis=1)

del df_demographic_train, df_demographic_test, features_train, features_test

## Convert to H20 Frame

start = time.time()
hf_test = h2o.H2OFrame(df_test)
end = time.time()
print(f"Converting test dataframe took {(np.round((end-start)/60, 2))} min")

start = time.time()
hf_train = h2o.H2OFrame(df_train)
end = time.time()
print(f"Converting train dataframe took {(np.round((end-start)/60, 2))} min")

## Prepare for H20 Models

hf_train['Label'] = hf_train['Label'].asfactor()
hf_test['Label'] = hf_test['Label'].asfactor()
response = "Label"
predictors = [predictor for predictor in hf_train.columns if predictor != 'Label']

train, valid = hf_train.split_frame(ratios=[.8], seed=1234)

## XGBoost Estimator

from h2o.estimators import H2OXGBoostEstimator
start = time.time()
xgb_model = H2OXGBoostEstimator(seed=1234)
xgb_model.train(x=predictors, y=response, training_frame=train, validation_frame=valid)
end = time.time()
print(f"Cell took {np.round((end-start)/60, 2)} min")

## Save Model

#h2o.save_model(model=xgb_model, path=dirPath + f"H20Models/XGBoost/_embed_{EMBED_LENGTH}", force=True)
h2o.save_model(model=xgb_model, path=dirPath + f"H20Models/XGBoost/combined_vists_embed_{EMBED_LENGTH}", force=True)

## Save Model Performance

train_perf = xgb_model.model_performance()
test_perf = xgb_model.model_performance(hf_test)

def saveResults(preformance, csvFile, embed_len, dataset='train', printResults=False):
    logLoss = preformance.logloss()
    mse = preformance.mse()
    auc = preformance.auc()
    aucpr = preformance.aucpr()
    recall = preformance.recall()[0][0]
    precision = preformance.precision()[0][0]
    specificity = preformance.specificity()[0][0]
    accuracy = preformance.accuracy()[0][1]
    
    if dataset == 'train':
        csvFile.write(f"{embed_len}, {accuracy}, {auc}, {aucpr}, {logLoss}, {mse}, {recall}, {precision}, {specificity}, ")
    elif dataset == 'test':
        csvFile.write(f"{accuracy}, {auc}, {aucpr}, {logLoss}, {mse}, {recall}, {precision}, {specificity}\n")
    else:
        raise ValueError(f'dataset can only be \'train\' or \'test\'. \'{dataset}\' is incorrect.')
        
    if printResults:
        print(f"{dataset} Log loss: {logLoss}")
        print(f"{dataset} MSE: {mse}")
        print(f"{dataset} AUC: {auc}")
        print(f"{dataset} AUCPR: {aucpr}")
        print(f"{dataset} Recall: {recall}")
        print(f"{dataset} Precision: {precision}")
        print(f"{dataset} Specificity: {specificity}")
        print(f"{dataset} Accuracy: {accuracy}")

## Save model perfromance

#csvSavePath = dirPath + "h2o_XGB_ModelResults.csv"
csvSavePath = dirPath + "h2o_XGB_Combined_Visits_Model_Results.csv"
saveFile = open(csvSavePath, 'a')
saveResults(train_perf, saveFile, EMBED_LENGTH, dataset='train')
saveResults(test_perf, saveFile, EMBED_LENGTH, dataset='test')
saveFile.close()

pd.read_csv(dirPath + "h2o_XGB_Model_Results.csv") # Visits treated independently

pd.read_csv(dirPath + "h2o_XGB_Combined_Visits_Model_Results.csv") # Visits combined

## Analyze Predictions

!ls Data/H20Models/XGBoost

xgb_model = h2o.load_model('/home/ubuntu/BioMedProject/Data/H20Models/XGBoost/combined_vists_embed_16/XGBoost_model_python_1629824610670_1')
pred = xgb_model.predict(hf_test)
true_values = df_test['Label'].values.tolist()
pred_values = h2o.as_list(pred['predict']).values.tolist()
pred_values = [val[0] for val in pred_values]
TP = 0
FP = 0
TN = 0
FN = 0
for i in range(len(true_values)):
    if true_values[i] == pred_values[i]:
        if true_values[i] == 0:
            TN += 1
        else:
            TP += 1
    else:
        if true_values[i] == 0:
            FP += 1
        else:
            FN += 1
total = TP + FP + TN + FN
print(f"True pos: {TP} ({TP / total})")
print(f"True neg: {TN} ({TN / total})")
print(f"False pos: {FP} ({FP / total})")
print(f"False neg: {FN} ({FN / total})")
